refactor: global package 역할별 분리
