+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import roomescape.controller.dto.ThemeRequest;
-import roomescape.global.AdminOnly;
+import roomescape.global.auth.AdminOnly;
 import roomescape.service.ThemeService;
 
 @AdminOnly
