+++ src/main/java/roomescape/global/auth/AdminOnly.java
@@ -1,4 +1,4 @@
-package roomescape.global;
+package roomescape.global.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
