+++ src/main/java/roomescape/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package roomescape.global;
+package roomescape.global.config;
 
 import java.util.List;
 
@@ -6,6 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import roomescape.global.auth.AdminAuthorizationInterceptor;
+import roomescape.global.auth.LoginMemberArgumentResolver;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
