+++ src/test/java/roomescape/controller/UserReservationControllerTest.java
@@ -31,9 +31,9 @@ import roomescape.domain.Member;
 import roomescape.domain.Role;
 import roomescape.domain.exception.DomainErrorCode;
 import roomescape.domain.exception.RoomescapeException;
-import roomescape.global.DomainErrorHttpMapper;
-import roomescape.global.LoginMemberArgumentResolver;
-import roomescape.global.WebConfig;
+import roomescape.global.exception.DomainErrorHttpMapper;
+import roomescape.global.auth.LoginMemberArgumentResolver;
+import roomescape.global.config.WebConfig;
 import roomescape.service.AuthService;
 import roomescape.service.ReservationService;
